net: phy: Add phy_interface_is_rgmii helper
commit e463d88c36d42211aa72ed76d32fb8bf37820ef1 branch ti-linux-4.4.y

RGMII interfaces come in 4 different flavors that the PHY library needs
to care about: regular RGMII (no delays), RGMII with either RX or TX
delay, and both. In order to avoid errors of checking only for one type
of RGMII interface and miss the 3 others, introduce a convenience
function which tests for all values.

Suggested-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>