ARM: dts: Add support for dra72-evm rev C (ES2.0)
commit 7f93c3879a57fd415fad8ed5d7ede8b183b34b03 branch ti-linux-4.4.y

DRA72-EVM now has an upgrade to Rev C with ES2.0 silicon. As part of
this change, a few updates were factored in that were software
incompatible with previous board in few areas:
- We now use DP83867 ethernet phy instead of older DP838865 which fails
  in certain usecases.
- Two ethernet ports now instead of the single one in rev B.
- polarities changed for certain pcf gpios
- Due to SoC phy current requirements, VDDA supplies are split between
 ldo5 and ldo2 (ldo2 was previously unused). NOTE: DSS (VDDA_VIDEO) is
 still supplied by ldo5.

NOTE: It does not make much sense to spin off a new board compatible
flag since there is no real benefit for the same.

Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>