OMAPDSS: DPI: Internally use the clock source ID instead of the PLL data
This means that struct dpi_data does not contains the PLL information
anymore but a clock source ID instead.
It allows us to transport the HSDIV information along with the PLL
information, and make the code more generic.
This is how it works in 4.4

The clock source ID is determined based on the platform type and the
channel used (just as PLLs were before). DRA7x platforms are handled
separately because the PLL mux must also be configured at this point.

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>