arm: dts: infoadas: Update GPU CMA memory allocations
dra7-evm and dra76-evm dts files reserve 256MB of CMA memory
from 0xC0000000 to enable the driver to allocate from CMA region.
This support was to demonstrate GPU allocations  to come from high
zone in LPAE configuration.

But visionSDK usecases on EVMs are tested for 1GB DDR available
for kernel(mem=1024M in bootArgs). So default GPU allocations are
not accessible.

Reserve a smaller size in a location which is accessible for 1GB
configuration.

Signed-off-by: Ramprasad N <x0038811@ti.com>